// src/models/server.ts
export interface PlexServer {
    accessToken: string; 
    name: string;
    address: string;
    port: string;
    version: string;
    scheme: string;
    host: string;
    localAddresses: string; 
    machineIdentifier: string;
    createdAt: string; 
    updatedAt: string;
    owned: string;
    synced: string;
}