// src/models/session/player.ts
export interface PlexSessionPlayer {
    address: string;
    device: string;
    machineIdentifier: string;
    model: string;
    platform: string;
    platformVersion: string;
    product: string;
    profile: string;
    remotePublicAddress: string;
    state: string;
    title: string;
    vendor: string;
    version: string;
    local: string;
    relayed: string;
    secure: string;
    userID: string 
}