// src/models/user.ts
export interface PlexUser {
    id: string;
    username: string;
    email: string;
    recommendationsPlaylistId: string;
    thumb: string;
    protected: string;
    home: string;
    allowTuners: string;
    allowSync: string;
    allowCameraUpload: string;
    allowChannels: string;
    allowSubtitleAdmin: string;
    filterAll: string;
    filterMovies: string;
    filterMusic: string;
    filterPhotos: string;
    filterTelevision: string;
    restricted: string;
}